Seperating nutmeg monitors into a seperate file
